Feature branch to test fix to publish workflow
Temurin just recently upgrade their latest Java 11 version to 11.0.20,
but that causes an error (bad scalamatics.jar file), so we need to lock
in the version to 11.0.19

Added a few mods in ci.yaml to allow the 'build assembly' step to run
just on the push of this feature branch, will remove those changes
before checkin.

Will change Java version in the next commit.