Lazy loading of tiles
Changes so tiles are read lazily in *Tile.scala when the first access is
made to the field of ForestChangeDiagnosticTile, rather than always
reading a relevant tile unconditionally in *GridSources.scala. Uses lazy
Scala fields in ForestChangeDiagnosticTile, whose values are computed
if/when they are accessed for the first time.

This is useful to avoid an unnecessary tile load if there a conditional
access to a tile in *Summary.scala/getGridVisitor.visit.

Added some temporary print statements in Layer.scala to show when tiles
are being fetched.