Change landmark_area to landmark_by_category_area
It looks like there will be multiple categories for indigenous/community
areas, so change landmark_area as a single number to
landmark_by_category_area to be map/object with zero or more categories
with areas. Indigenous areas in Argentina are not classified, so they
use the default category of "Not Reported".