GTC-2622 Add test for gfwpro_dashboard
To allow for the test, I needed to change GfwProDashboardAnalysis to
return the RDD, instead of doing the full export and returning nothing.
(Similar to ForestChangeDiagnosticAnalysis, which also returns an RDD so
that writing a test is possible.)

I added an input file dash.tsv for the test. I also added a file
idn1_5Gadm.tsv which is just the row from gadm36_adm2_1_1.tsv for
IDN1.5, which is the only area needed for the particular input geometry
in dash.tsv.

In this next commit, I will add a pinning option, so the output stays
constant even as new integrated alerts datasets are added.