Add ability to pin a version for a dataset.
The pinned command-line option already exists, but it was no longer
functional, so I re-enabled it. To do that, I had to add a new 'pinned'
argument to GfwConfig.get(), so the 'pinned' value could control which
version is used from the catalog.

We are using it specifically for the gfwpro_dashboard test, where we
need to pin the integrated_alerts version so that the dashboard results
from the test are constant.