GTC-2641 Add the boot script that we use for EMR nodes
This script has not been under source code control before. As the fix
for GTC-2641, I added a new line loading a specific version of libcurl
(version 7.76.1). Otherwise, we always get the newest version of libcurl
(currently version 8.5), and that is causing the issue (socket
disconnects on repeated access to S3). I'm not yet sure if it is a bug
in the newer versions of libcurl or an intended feature.