Add in the country-specific wdpa and primaryForest columns.
We decided that the country-specific columns will include the wdpa and
primary forest intersection, just like prodes has, especially since the
country-specific columns will soon replace the prodes columns. For
Argentina, the UI may or may not show the wdpa / primary forest data,
but it will be available.

Add back in checking braBiomesPresence, so we are more likely to catch
any locations that contain area in both Argentina and Brazil.