GTC-2732 Add JRC map intersection to AFi analysis
The native JRC map is close to 10m resolution, so we move the grid and
all the other datasets to 10m resolution. JRC is the more official
forest map for EUDR vs. the SBTN natural forests map. Added two new
columns in the output, jrc_forest_cover__extent and
jrc_forest_cover_loss__ha, in analogy to the natural_forest_* columns.