GTC-2747 New columns in AFi which is forest loss per year since 2021
Added two new columns which are forest loss for each year from 2021 for
both the natural forest (SBTN) and the JRC forest cover in each location.

AFi was previously written mainly using data frames. This worked well,
because there were only scalar numbers in each column - no complex data
structure like per-year forest loss
(ForestChangeDiagnosticDataLossYearly). With this change, I need to
aggregate on two new values which are per-year-forest loss. This is not
easy to do in dataframes, because you have to define user-defined
aggregate functions (and possibly user-defined types). And I actually
ran into a run-time bug compiling the generated Java code for the needed
user-defined aggregate function, which I couldn't figure out how to work
around.

So, I switched the code to do most of the aggregation in RDDs, similar
to forest_change_diagnostic. This means we don't have to define
user-defined aggregation functions, and in fact, can use the normal
merge function of ForestChangeDiagnosticDataLossYearly.

I defined the range of years for forest loss in
AfiCommand.{TreeCoverLossYearStart,TreeCoverLossYearEnd}. Added some
extra args for the empty() and prefilled() methods of
ForestChangeDiagnosticDataLossYearly, so it works for different year
ranges. I set the end year to 2023 already, since TCL 2023 is coming in
at about the same time.

Removed a bunch of unused imports in various files.