GTC-2734 Add JRC/SBTN, confidence levels, and total_ha to integrated alerts.
For daily alerts, add new columns that divide the daily alerts into a
3x3 cross-product of (whole area, SBTN forest only, JRC forest only) x
(lo, med, hi) alert confidence

Also, add a new 'total_ha' column which is the area of the location.
This area will be used by the front-end to calculate measures such as
"daily alerts per hectare".

Currently, I am calculating the entire location area in the usual
per-pixel way with Geodesy.pixelArea. We were already calculating the
area per pixel whenever isTreeCoverExtent30 was true (more than 30% tree
cover). From a look at some input sets, it seems like this is usually a
large part of the location areas. So, doing the area calculation for the
entire geometry shouldn't be much more expensive.

However, I will make sure that the isTreeCoverExtent30 area is a large
part of the whole area for most datasets, once this change is in. If it
is not, then it may give better performance to calculate the total area
of locations outside of the pixel function using an equal-area
projection (for which I already have some working code).