GTC-2651 (part 2) - handle "antarctica case" for dashboard as well
This is handling GTC-2651 for gfwpro_dashboard as well. If all locations
in the input list are outside of the TCL extent (e.g. all in Antartica),
then the first analysis will fail because there is no boundary that
spatial join want to access. So, exit out of the fireStats if the RDD is
completely empty, because all locations are outside the TCL extent.