Reuse analysis results for same full-window geometries from different locs
This is an optimization that Justin did a while ago for GFW in
SummaryRDD.scala. I'm porting it to ErrorSummaryRDD.scala, which is what
GFWPro uses.

If several locations overlap an entire window, then only do the analysis
on that window once, and use it as one of the partial results for all
those locations. This optimization is especially useful when locations
have a lot of overlap, possibly because there are lots of large,
overlapping, supply sheds, or because of the batching of lots of
related lists in the ALERTS-Batch analysis.