Remove code in FCD related to geometry diffs and intermediate results
We've been ignoring geometry diffs (location id == -2) for a long time,
and we pass in any intermediate results.

By removing this code, we avoid computing the centroid of each location
geometry, inserting it into the feature id, then removing it from the
feature id without having used it at all.

The test change was just a change in order of categories within a single
result, which is no actual change of the results.