GTC-2913 Switch to new Landmark version v202408
We created a new dataset gfw_indigenous_community_and_indicative_lands
(version v202408) that merges in the indicative lands that have been
separated out to a different Landmark dataset in the most recent
Landmark releases (v202407). We use the combined dataset in our analysis
and use its tile cache for display in Pro, so Indicative lands can be on
the same map as the Indigenous/Community lands.