GTC-3077 Give NoIntersectionError in AFi if geometry too small
In the AFi analysis, we currently don't return any row at all for a
location whose geometry does not intersect the centroid of any pixel
(mostly because the geometry is small compared to the pixel size). Add
similar code as in FCD to detect this case and return
NoIntersectionError.

Same change for GFWProDashboard as well.

Added a NoIntersection test for AFi.