GTC-3055 Added new GHG (green-house gas) analysis
This analysis computes the DLUC (direct land use change) GHG emissions
factors for each location, based on the tree loss information for the
location and the specified yield or commodity (which gives a default
yield value). The emissions are calculated by a 20-year discounted
formula, as described in GHGRawDataGroup.scala. We compute the emissions
factors for each year in 2020-2023.

The new set of GHG Analysis files are most closely modeled on the
ForestChangeDiagnostic analysis files, but are simplified as much as
possible. The GHGSummary file has the logic that computes the crop yield
for a particular pixel location, including making use of the primary
crop yield datasets or the backup yield CSV file. It also computes the
emissions due to tree loss. The GHGRawDataGroup file does the emissions
factor computation.

Here are some of the other needed changes:
 - Added gross emissions datasets (already in use for some carbon flux analyses) to
   the Pro dataset catalog.
 - Added a new dataset per commodity (currently 6 commodities, will add
   more later) that gives the average crop yield in each 10km area. Added
   a generic MapspamYield layer to access these commodity datasets.
 - Added a CSV file with "backup" crop yields per GADM2 area, when the
   primary crop yield rasters don't have a value for a particular pixel.
   This CSV file (which is < 26 Mbytes) is broadcast to all nodes (not
   each task), so it is available for lookup, and avoids Spark tasks
   swamping the Data API with requests. The CSV file is specified by
   command-line option and will be placed on S3.
 - Added a new Feature and FeatureId "gfwpro_ext" that includes
   "commodity" and "yield" fields, which are needed to specify the
   exact crop yield or commodity grown at each location.
 - Changed ErrorSummaryRDD to allow passing in the featureId to the
   polygonalSummary code as part of kwargs. This is needed so we can
   pass the yield and commodity information into the GHGSummary code. We
   need the yield/commodity during the per-pixel analysis.
 - Added a GHG test. The input file for the test case (ghg.tsv) tests a
   variety of commodities and yields, and even the error case where a
   default yield for a commodity is not found at all for a location. A
   partial backup GADM2 yield file is includes in the test files
   (part_yield_spam_gadm2.csv)