Change reading of backup yield file.
The method via spark.read.csv() does not work. It's seems like to be a
bug (maybe with our older version of Spark), where reading the CSV to a
DataFrame and then immediately collecting it back to the master fails
because the Spark connections are not fully enabled yet.

Changed to reading the file on S3 or on the local file system more explicitly.