Skip using GDAL if GFW_SKIP_GDAL env variable is set.
This is using the method that Justin used when he didn't have the right
GDAL version. We avoid using GDAL by using GeoTiffRasterSource, rather
than GDALRasterSource in loading tiffs in Layer.scala. This is useful
for local runs where the user doesn't have the right GDAL version, but
results in much slower runs (about 1.8x slower).

If we are skipping the use of GDAL, I automatically replace any
'gdal-geotiff' folder in the URI with 'geotiff', since we won't be able
to deal with special GDAL-optimized tiffs. This is a fairly special
case, but it's only for local use where people don't want to set up the
correct GDAL version.