GTC-3143 Modify Flagship IntegratedAlerts analysis to produce integrated disturbance alerts values (#257)
* Changing IntegratedAlerts analysis to all produce int-dist alert results

We compute int-dist alerts from the base alerts and dist alerts, so this
analysis can run in parallel with the job that actually creates the new
int-dist raster tile set.

* Remove unused climateMask variable in IntegratedAlertsDataGroup.

Also, put in comments of various layers that are broken (always false),
especially for the 10/100000 grid.

* Replace integrated alerts columns by integrated dist alerts.

* Don't put integrated alerts fields in IntegratedAlertsDataGroup

Changes based on Justin's comments.

Don't want the in-memory data to multiply by having both integrated
alerts and integrated dist alerts in the DataGroup. Also, added some
clarifying comments.

* Add TreeCover2022 layer, add it as column in IntegratedAlerts analysis

tree_cover_2022 is defined as:

(umd_tree_cover_density_2010/v1.6 >= 30) and
(umd_tree_cover_height_2020/v2022 >= 3) and
(umd_tree_cover_loss/v1.12 != 2021)

It is the tree cover filter that we use as an option for dist alerts and
(soon) integrated-dist alerts.

* Convert all relevant integrated alerts layers to use gdal-geotiff

* Add comment for DistDateConfLayer